+++ CHANGELOG.md
@@ -1,7 +1,9 @@
 ## 1.1
 
 - Use lazy val to initialize SecureRandom, so that initialization occurs also async
-- Refactor authenticators and add HeaderAuthenticator
+- Refactor authenticators and add HeaderAuthenticator and SessionAuthenticator
+- Split up SocialProvider.authenticate method into authenticate and retrieveProfile methods
+- Remove authInfo from SocialProfile
 
 ## 1.0 (2014-06-12)